Port r2 to Python 3.11 runtime: stdlib imports, idiom semantics, py3-only modules
Second layer on top of the syntax pass (4cf4a54). The tree already parsed; every
change here is about what happens at *runtime* on py3.

Stdlib imports retargeted (per-site, not blind):
- urlparse -> urllib.parse (alias; all 6 used attributes exist there)
- cStringIO/StringIO -> io, choosing BytesIO for binary producers
  (image.save, gzip.GzipFile, zipfile, HTTP response body) and StringIO for
  text (csv.writer, tokenize, readfp, lxml) -- 12 sites decided individually
- urllib2 -> urllib.request, EXCEPT URLError/HTTPError which live in
  urllib.error on py3 (they are not in urllib.request)
- httplib -> http.client; ConfigParser -> configparser (SafeConfigParser was
  removed in 3.12, aliased to ConfigParser); Queue -> queue;
  Cookie -> http.cookies (only the CookieError import; the r2 Cookie class is
  unrelated and untouched); HTMLParser -> html.parser; cPickle -> pickle;
  new.classobj -> type()
- bare 'import urllib' + urllib.quote/quote_plus/unquote/unquote_plus/urlencode
  -> urllib.parse.* (py2's top-level urllib functions moved)

Idiom semantics:
- .iteritems/.itervalues/.iterkeys -> items/values/keys (316 sites, 114 files);
  the one custom def (zookeeper LiveConfig) renamed to items() so its
  also-renamed callers resolve
- xrange -> range; basestring -> str; long() -> int(); isinstance(..., long)

Text handling:
- rewrote r2/r2/lib/unicode.py: _force_unicode decodes bytes (utf-8, latin1
  fallback) and str()s anything else; _force_utf8 no longer wraps encode() in
  str() -- that would have produced "b'...'" on py3
- converted the 19 unicode(...) call sites; isinstance(x, unicode) -> str;
  class _Unsafe(unicode) -> (str); no __unicode__ methods existed

Class/ordering:
- __metaclass__ = M -> metaclass= in the class statement (9 classes)
- py2 __cmp__ (ignored on py3) -> __lt__/__eq__/__hash__; sort(cmp=f) ->
  sort(key=functools.cmp_to_key(f)); cmp(a,b) -> (a>b)-(a<b)

Removed-in-py3:
- inspect.getargspec -> getfullargspec; time.clock -> time.perf_counter;
  reload() -> importlib.reload(); itertools.izip/ifilter -> zip/filter;
  string.letters/lowercase/uppercase -> string.ascii_*;
  old-style `raise "string"` -> raise ValueError(...)

Verified
- 0 of 314 files fail a parse under Python 3.11.15 and 3.14.7; compileall clean
- pyflakes: py2-builtin undefined-name findings 26 -> 19, the 7 removed being
  exactly basestring/cmp/execfile/long/reload/unicode/xrange; ZERO introduced

NOT addressed (separate work)
- dependency layer: Pylons 1.0.1 (174 files), pycassa (26), SQLAlchemy 0.8.4
  (16), boto2, amqplib; C extensions (6 .pyx, Cfilters, snudown)
- semantic risks that need behavioural tests, not static ones: int/int division
  (/ changed meaning), map/filter/zip returning iterators, dict view mutation
  during iteration, and str-vs-bytes in code paths pyflakes cannot see

+++ r2/r2/config/feature/state.py
@@ -123,7 +123,7 @@ class FeatureState(object):
         # Mix the feature name in with the seed so the same users don't get
         # selected for ramp-ups for every feature.
         hashed = hashlib.sha1(self.name + seed)
-        bucket = long(hashed.hexdigest(), 16) % self.NUM_BUCKETS
+        bucket = int(hashed.hexdigest(), 16) % self.NUM_BUCKETS
         return bucket
 
     @classmethod
